[루틴] 캘린더 이벤트 클릭시 루틴정보 get, post api작성 완료, [AWS] 루틴테이블 기본키 수정

+++ server/models_aws/Schedule.js
@@ -3,7 +3,7 @@
 module.exports = function(sequelize, DataTypes) {
   return sequelize.define('Schedule', {
     ScheduleId: {
-      type: DataTypes.INTEGER(11),
+      type: DataTypes.BIGINT,
       allowNull: false,
       primaryKey: true,
       autoIncrement: true
